% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_tiff_3x7.R
\name{bar_tiff_3x7}
\alias{bar_tiff_3x7}
\title{Create a 3 x 7 Tiff Image}
\usage{
bar_tiff_3x7(df, x, y, t, xlab = x_nm, ylab = y_nm)
}
\arguments{
\item{df}{The dataframe loaded into the current environment to be used
for the image creation.}

\item{x}{The categorical variable to be used for the x-axis of the graph}

\item{y}{The discrete variable to be used for the y-axis of the graph}

\item{t}{The title desired for the graph (in quotes)}

\item{xlab}{The label for the x-axis if different from the variable name in
dataframe. By default, it will use the variable name.}

\item{ylab}{The label for the y-axis if different from the variable name in
dataframe. By default, it will use the variable name.}
}
\description{
This function has been designed to provide a standardized format for
all FDA BEST intiative graphs. It will generate a 3x7 Tiff image of
high resolution (300 dpi), and will save it to the working directory.
}
\author{
David Carnahan <dcarnahan@us.imshealth.com>
}
