% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdabest.R
\docType{package}
\name{fdabest}
\alias{fdabest}
\alias{fdabest-package}
\title{A package for creating standardized graphical and tabular information for the
FDA BEST intiative.}
\description{
This package provides standardized formatting for the most common sizes of high
resolution graphics and tables used for disseminating information to the FDA in
support of the BEST intiative.
}
\author{
David Carnahan \email{dcarnahan@us.imshealth.com}
}
